+++ native/app/components/ListPanel.js
@@ -16,18 +16,8 @@ import { trackEvent } from '../utils/metrics';
 
 import ListPanelEmpty from './ListPanelEmpty';
 import ListPanelLoading from './ListPanelLoading';
-import Snackbar from './Snackbar';
 
-const SNACKBAR_ANIMATION_DURATION = 250;
-const SNACKBAR_HEIGHT = 48;
-
-const SYNCED_SNACKBAR = {
-  text: 'Notes synced!',
-  color: COLOR_DARK_SYNC,
-  action: null,
-  onDismiss: null,
-  duration: 3000
-};
+var _ = require('lodash')
 
 class ListPanel extends React.Component {
   constructor(props) {
@@ -47,7 +37,7 @@ class ListPanel extends React.Component {
 
     this._onRefresh = () => {
       if (this.props.state.sync.isConnected === false) {
-        ToastAndroid.show('You are offline.', ToastAndroid.LONG);
+        ToastAndroid.show(browser.i18n.getMessage('toastOffline'), ToastAndroid.LONG);
       } else {
         trackEvent('webext-button-authenticate');
         this.setState({ refreshing: true });
@@ -56,6 +46,10 @@ class ListPanel extends React.Component {
         }).catch(() => {
           this.setState({ refreshing: false });
         });
+
+        setTimeout(() => {
+          this.setState({ refreshing: false });
+        }, 12000)
       }
     }
 
@@ -91,83 +85,6 @@ class ListPanel extends React.Component {
       }
     };
 
-    this._showSnackbar = (snackbar) => {
-      if (!this.state.snackbar && snackbar) {
-        this.setState({
-          snackbar,
-          snackbarVisible: true,
-        });
-
-        Animated.timing(this.state.fabPositionAnimation, {
-          toValue: 1,
-          duration: SNACKBAR_ANIMATION_DURATION,
-          useNativeDriver: true,
-        }).start();
-      } else if (snackbar && snackbar.color !== COLOR_DARK_WARNING &&
-        this.state.snackbar.text !== snackbar.text) {
-        this.snackbarList.push(snackbar);
-      }
-    };
-
-    this._hideSnackbar = () => {
-
-      return new Promise((resolve) => {
-        this.setState({
-          snackbarVisible: false
-        });
-
-        Animated.timing(this.state.fabPositionAnimation, {
-          toValue: 0,
-          duration: SNACKBAR_ANIMATION_DURATION,
-          useNativeDriver: true,
-        }).start(() => {
-          this.setState({
-            snackbar: null
-          });
-          if (this.snackbarList.length > 0) {
-            this._showSnackbar(this.snackbarList.shift());
-          }
-          resolve();
-        });
-      });
-
-    };
-
-    this._undoDelete = (deletedNote) => {
-
-      if (!this.ignoreUndo) {
-        this.ignoreUndo = true;
-
-        const promises = [];
-
-        if (deletedNote && Array.isArray(deletedNote)) {
-          deletedNote.forEach((note) => {
-            promises.push(props.dispatch(createNote(note)));
-          });
-        }
-
-        Promise.all(promises).then(this._hideSnackbar, this._hideSnackbar).then(() => {
-          this.ignoreUndo = false;
-        });
-      }
-    };
-
-    this._requestReconnect = () => {
-      if (!this.props.state.sync.loginDetails) {
-
-        this.props.dispatch(openingLogin());
-        this.snackbarList = [];
-        return Promise.resolve()
-        .then(() => fxaUtils.launchOAuthKeyFlow())
-        .then((loginDetails) => {
-          trackEvent('login-success');
-          this.props.dispatch(authenticate(loginDetails));
-          this.props.dispatch(kintoLoad());
-        }).catch((exception) => {
-          this.props.dispatch(reconnectSync());
-        });
-      }
-    };
   }
 
   componentDidMount() {
@@ -180,170 +97,29 @@ class ListPanel extends React.Component {
     NetInfo.removeEventListener('connectionChange', this._handleNetworkStateChange);
   }
 
-  componentWillReceiveProps(newProps) {
-    if (newProps.navigation.isFocused()) {
-      // Display synced note snackbar
-      if (this.props.state.sync.isSyncing &&
-          !newProps.state.sync.isSyncing &&
-          !newProps.state.sync.error &&
-          newProps.state.sync.isConnected !== false &&
-          this.state.appState === 'active' &&
-          newProps.state.sync.loginDetails) {
+  shouldComponentUpdate(nextProps) {
+    let shouldUpdate = false;
 
-        if (this.props.state.sync.isSyncingFrom === 'drawer') {
-          setTimeout(() => this._showSnackbar(SYNCED_SNACKBAR), 400);
-        } else if (this.props.state.sync.isSyncingFrom !== 'deleteNote') {
-          this._showSnackbar(SYNCED_SNACKBAR);
-        }
-      } else if (!this.props.state.sync.error && newProps.state.sync.error && newProps.state.sync.loginDetails) {
-        this._showSnackbar({
-          text: newProps.state.sync.error,
-          color: COLOR_DARK_WARNING,
-          action: null,
-          duration: 3000
-        });
-      } else if (!newProps.state.sync.loginDetails &&
-                 !newProps.state.sync.isOpeningLogin &&
-                 !newProps.state.sync.isPleaseLogin) {
-        this._showSnackbar({
-          text: newProps.state.sync.error,
-          color: COLOR_DARK_WARNING,
-          onPress: this._requestReconnect,
-          duration: 0
-        });
-      }
+    const changedProps = _.reduce(this.props.state, function (result, value, key) {
+      return _.isEqual(value, nextProps.state[key])
+        ? result
+        : result.concat(key)
+    }, []);
 
-      // If user login and reconnectSync snackbar is open
-      if (this.state.snackbar &&
-          this.state.snackbar.duration === 0 &&
-          this.state.snackbar.color === COLOR_DARK_WARNING) {
-          if (!this.props.state.sync.loginDetails && newProps.state.sync.loginDetails) {
-            this._hideSnackbar();
-          } else if (newProps.state.sync.isConnected === false) {
-            this._hideSnackbar();
-          } else if (newProps.state.sync.isOpeningLogin || newProps.state.sync.isPleaseLogin) {
-            this._hideSnackbar();
-          }
-      }
-
-      // Display deleted note snackbar
-      if (newProps.navigation.getParam('deletedNote')) {
-        // We store deletedNote to be able to recreate it if user click undo
-        const deletedNote = newProps.navigation.getParam('deletedNote');
-
-        // Erase params for future componentWillReceiveProps events
-        newProps.navigation.setParams({ deletedNote: null });
-
-        // Show snackbar
-        this._showSnackbar({
-          text: 'Notes deleted.',
-          color: COLOR_NOTES_BLUE,
-          action: {
-            text: 'UNDO',
-            onPress: () => {
-              this._undoDelete(deletedNote);
-            }
-          },
-          duration: 6000
-        });
-      }
-
-      if (!this.props.state.sync.selected && newProps.state.sync.selected) {
-        Animated.timing(this.state.fabOpacityAnimation).stop();
-        Animated.timing(this.state.fabOpacityAnimation, {
-          toValue: 0,
-            duration: 150,
-            useNativeDriver: true,
-          }).start(({ finished }) => {
-            if (finished) {
-              this.setState({
-                hideFab: true,
-                fabOpacityAnimation: new Animated.Value(0)
-              });
-            }
-          });
-      } else if (this.props.state.sync.selected && !newProps.state.sync.selected) {
-         this.setState({ hideFab: false });
-         Animated.timing(this.state.fabOpacityAnimation).stop();
-         Animated.timing(this.state.fabOpacityAnimation, {
-            toValue: 1,
-            duration: 150,
-            useNativeDriver: true,
-          }).start();
-      }
-    } else {
-      this.snackbarList = [];
+    // only render when notes collection changes
+    if (changedProps.length > 0 && _.intersection(changedProps, ['notes', 'appUpdates', 'kinto']).length > 0) {
+      shouldUpdate = true;
     }
+
+    return shouldUpdate;
+  }
+
+  componentWillReceiveProps(newProps) {
+
   }
 
   render() {
-    return (
-      <View style={{ flex: 1, position: 'relative' }}>
-        { this.renderList() }
-        <Snackbar
-          style={{
-            backgroundColor: this.state.snackbar ? this.state.snackbar.color : COLOR_DARK_SYNC,
-            position: 'absolute',
-            bottom: -1 * SNACKBAR_HEIGHT,
-            transform: [
-              {
-                translateY: this.state.fabPositionAnimation.interpolate({
-                    inputRange: [0, 1],
-                    outputRange: [0, -1 * SNACKBAR_HEIGHT]
-                  }),
-              },
-            ],
-          }}
-          visible={ this.state.snackbarVisible }
-          action={ this.state.snackbar ? this.state.snackbar.action : null }
-          theme={{ colors: { accent: 'white' }}}
-          onDismiss={(onPress) => {
-            this._hideSnackbar();
-            if (this.state.snackbar && this.state.snackbar.onPress && onPress) {
-              setTimeout(this.state.snackbar.onPress, 10);
-            }
-          }}
-          duration={ this.state.snackbar ? this.state.snackbar.duration : 3000 }
-        >
-          { this.state.snackbar ? this.state.snackbar.text : '' }
-        </Snackbar>
-
-        { this.props.state.kinto.isLoaded && !this.state.hideFab ?
-          <Animated.View style={[
-            {
-              position: 'absolute',
-              bottom: 0,
-              right: 0,
-              width: 92,
-              height: 84,
-              opacity: this.state.fabOpacityAnimation
-            },
-            {
-              transform: [
-                {
-                  translateY: this.state.fabPositionAnimation.interpolate({
-                    inputRange: [0, 1],
-                    outputRange: [0, -1 * SNACKBAR_HEIGHT]
-                  }),
-                },
-                {
-                  scale: this.state.fabOpacityAnimation.interpolate({
-                    inputRange: [0, 1],
-                    outputRange: [0.9, 1]
-                  }),
-                }
-              ],
-            }]}>
-            <FAB
-              small
-              color={COLOR_NOTES_WHITE}
-              style={styles.fab}
-              icon="add"
-              onPress={() => this.props.navigation.navigate('EditorPanel', { note: null }) }
-            />
-          </Animated.View> : null }
-      </View>
-    );
+    return this.renderList();
   }
 
   renderList() {
@@ -369,7 +145,11 @@ class ListPanel extends React.Component {
       return (
         <FlatList
           contentContainerStyle={styleList}
-          data={this.props.state.notes.sort((a, b) => { return a.lastModified <= b.lastModified ? 1 : -1 })}
+          data={this.props.state.notes.sort((a, b) => {
+            const aStamp = new Date(a.lastModified).getTime();
+            const bStamp = new Date(b.lastModified).getTime();
+            return aStamp <= bStamp ? 1 : -1
+          })}
           refreshControl={
             <RefreshControl
               refreshing={this.state.refreshing}