+++ native/app/utils/utils.js
@@ -1,16 +1,16 @@
-
 import striptags from 'striptags';
 
 function firstLine(content) {
-	content = content.split('</')[0].replace(/<\/p>|<\/li>/gi, '&nbsp;');
-	return striptags(content.replace(/&nbsp;/gi, ' ')).trim().substr(0, 250).replace(/&amp;/g, '&');
+  content = striptags(content.replace(/></gi, '>\n<')).replace(/&nbsp;/gi, ' ').trim().split('\n')[0];
+  if (!content) return '';
+  return content.substr(0, 250).replace(/&amp;/g, '&');
 }
 
 function secondLine(content) {
-	// Remove first line
-	content = content.replace(content.split('</')[0], '');
-	content = content.replace(/<\/p>|<\/li>/gi, '&nbsp;');
-	return striptags(content.replace(/&nbsp;/gi, ' ')).trim().substr(0, 250).replace(/&amp;/g, '&');
+  // Remove first line
+  content = striptags(content.replace(/></gi, '>\n<')).replace(/&nbsp;/gi, ' ').trim().split('\n')[1];
+  if (!content) return '';
+  return content.substr(0, 250).replace(/&amp;/g, '&');
 }
 
 module.exports = { firstLine, secondLine };